requirements.txt
Flask
google-generativeai
openai
python-dotenv
